+++ db/migrate/20260912000000_add_type_to_omen_readings.rb
@@ -1,0 +1,6 @@
+class AddTypeToOmenReadings < ActiveRecord::Migration[8.1]
+  def change
+    add_column :omen_readings, :type, :string
+    add_index :omen_readings, :type
+  end
+end